refactor excution code to be more maintainable
